Re-implement compiler with AST and IR (#1)
* Handle conditionals

- add modules to support multiple statements/expressions
- add gotos and labels to IR
- update codegen to handle symbolic addresses

TODO: create multiple basic blocks

* Construct CFG with BBs for if/else statement

* Implement reverse post order to traverse CFG

* Generate code by traversing CFG

* Fix non-popping of conditional value

* Change postOrderTraverse argument

* Get global variables initially working

* Implement global variable assignment

* Add consts with static error checking

Need to come up with a better way of handling errors when checking the AST.

* Switch back to using Tokens for variable AST

We will want to show line numbers in error messages, so copying the
Token preserves this information.

Need to separate error reporting from the parser so that later stages
can report errors too. Need to refactor out several components in
compiler.c and implement symbol table.

* Create separate semantic check with error handling

* Don't bail early in case of error

Better to show as many as possible to user.

* Create block statement

* Proper variables implementation with symbol table

A much improved implementation of variables that properly separates
concerns. Variables are stored in symbol table and all semantic checks
performed on that and the AST.

Most things from the CI compiler (e.g. putting locals in array) is moved
to codegen as it is backend-dependent.

* Support functions and arguments

* Separate Scope and ExecutionScope

* WIP upvalues

* Get nested functions to work correctly

* Maybe get functions working?

Change order in which CFGs are evaluated to follow execution order.
Add ExpressionStatement so that we can add a POP opcode afterwards.

* Add remaining binary expressions

* Implement while loops

* Implement for loops

Need to test optional clauses work ok.

* Make optional clauses work properly

for (;;) { } does an unnecessary pop. Need to investigate but low
priority

* Support AND

* Support OR

* Get basic class and method functionality working

* Add support for 'this'

* Support 'this' properly

* WIP: getting inheritance to work

* Set up test runners

* Rename test dirs

* Fix closure test bugs

* Fix some return stmt and variable errors

* WIP: reverting closure optimisation

* Fix invalid line numbers in runtime errors

* Pass Token around for line numbers

* Remove symbol lookup where it causes errors

Need a more sophisticated way to look up methods etc in class scope,
but need to know their type. Just using strings from token for now.

* Comment out unsupported minus minus

* Fix tests expecting error on line

* Improve label iteration logic. OP_CLOSURE todo

* Handle empty for statements

* Fix test assertions

* Only check current scope for methods

* Fix incorrect closure handling

* Hopefully fix scope errors in for and while

* Fix codegen errors in for/while

* Use OP_FALSE/TRUE literals

* Allow variable shadowing

* Bring back shadowing tests

* Fix set property error and update test assertions

* Add function name to locals

* Update test expectations

* Update test expectations now that shadowing is allowed

These tests fail because the semantic check sees the inner variable
and makes the symbol local, but at runtime the local variable does not
yet exist and so we fail to find anything when resolving local.

How to tell it to resolve a global?

* Prevent undefined local variables shadowing globals [HACK]

* Get basic version of inheritance working for globals

* Bring back inheritance-related tests

* WIP: classes share context and so are broken

* Move classes to separate context

* Fix remaining issues with inheritance. AST basically done!

The test suite is passing except for:
- doing a first semantic check for functions so that a function can
reference functions that haven't been seen yet.
- adding a check that `super` is not used in classes that do not have a
defined base class.
- built-in functions need to be added to semantic checks.

Next steps:
- add in the missing things in compiler.c such as postfix ++/--, arrays.
- lots of refactoring and tidy up code.
- try to make WorkUnits, ExecutionContexts etc less confusing.
- delete C test suites, tidy up repo.